module: network: add UI part of module for static ip setup
This module offers a page to enter the following details after the ssh
key selection page:
- network interface name,
- IP address,
- mask,
- gateway.

For now, the network interface name field is not checked for validity.
This is because, it will be reworked to show the list of available
interfaces. Currently, this provides a simple text box to enter the text
with a placeholder value en3ps0. Also, one must enter it to go forward.

The IP address text box is validated, so that user can only enter a
value in the '0.0.0.0' format, where integers can have values from 0 to
255. This is also a mandatory entry.

The Mask has a default value of 24. And this text box can be kept empty.
Also the user entry is validated to integers from 0 to 32.

The gateway text box behaves exactly like IPv4 text box.

If an input is valid, a green tick icon is shown near the text box.
Otherwise, it shows a red cross icon with a message showing a hint.
The icons used are inside the images folder

Note that this is to be used for any network related customization that
is why a generic module name. Also, this module is created following the
example of users module in Calamares.

Signed-off-by: Aiswarya Kaitheri Kandoth <aiswarya.kandoth@savoirfairelinux.com>