imageselection: support non-JSON metadata image files
JSON files can be added to each SEAPATH images to provide metadata
about the image (name, version, description). However, it is not mandatory
to have a JSON file for each image (particularly for custom images).
This commit adds support for the installer to also detect image files
without a JSON metadata file.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>