modules: networkcfg: mount: fix network config installation for debian
This fixes the creation of '.network' config file for debian.

The network config for debian was not working, because the
etc_mount_point is only for yocto flavor. Hence, this adds the rootfs
mount point of debian to global storage as rootMountPoint. Then it
constructs the path to etc and assign a value for etc_mount_point.

Signed-off-by: Aiswarya Kaitheri Kandoth <aiswarya.kandoth@savoirfairelinux.com>