modules/rawimage: extend persistent partition
Based on the behavior of the original SEAPATH flasher [1].
Original commit message:
Currently, the size of the persistent partition is set by the WIC file. The
problem is if we want a bigger partition size, it will make final WIC
image bigger too, even if it contains no data.
To resolve this, this commit extends during flash the size of the
persistent partition to maximum freespace available on the disk.

[1]: https://github.com/seapath/meta-seapath/commit/0041df3f2264033dc4c1ed55374a6db133d3d94e

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>