modules/imageselection: retrieve partition layout from the selected image
Each time an image is selected in the imageselection module, we now
parse its GPT headers to extract the partition layout. This layout is
stored in the GlobalStorage under the key "gptPartitions", and will be
used later by the partition module to preview the partitions to be
created on the device.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>