modules/partition: add previewCreatePartition
Currently, the PartitionCoreModule lacks a method to preview the
creation of a partition in the in-memory model without actually
scheduling any jobs (e.g format and create partition).
This commit adds the method previewCreatePartition() to the
PartitionCoreModule class, which allows inserting a new Partition
object into the partition model for a given Device, without creating
any jobs.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>