modules/partition: add partition preview
Currently, partition viewpage only show a blank page. Commit [1]
removed the Calamares partition preview as it was based on its own
partition layout and not the one installed by the SEAPATH image.

This commit introduces a new partition preview based on the partition
layout extracted from the SEAPATH image. Filesystems and partitions are
dummy created in the preview to reflect the layout defined in the
image, but no actual formatting or partitioning jobs are created at this
stage. The preview is refreshed each time the user navigates to the
partitioning page, to reflect any changes made by the user, like
selecting another image.

[1]: 03ef50d42af12459f266b30c82452bf0655968a4

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>