modules/rawimage: wait for partition to be created
Currently, when extending the persistent partition, the e2fsck command
is executed right after the parted command. In some cases, the later
can fail as the kernel still didn't updated the device with the new
partition informations. As a result, an exception is generated, and the
module fails.

To fix this commit:
    * Force the kernel to be inform of the latest partition changes
    * Wait for udev rules to be applied
    * Wait for the device partition to appear

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>